https://online.sbis.ru/opendoc.html?guid=9d605b23-d688-4e39-9d61-b7ea6c999494 IConsole
