+++ src/Application/_State/StateReceiver.ts
@@ -2,7 +2,7 @@
 import { IStateReceiver, IStateReceiverMeta } from 'Application/Interface';
 import type { ISerializableState } from 'Application/_State/Interface';
 
-import { IConsole } from 'Application/_Interface/IConsole';
+import type { IConsole } from 'Application/_Env/IConsole';
 
 /**
  * @author Санников К.
